web/requirements.txt
flask==1.1.2
flask_login==0.5.0
Flask-Migrate==2.7.0
flask_wtf==0.14.3
flask_sqlalchemy==2.4.4
sqlalchemy==1.3.23
gunicorn==0.17.4
flask-cors==2.1.3
Flask-JWT-Extended==3.25.0
Flask-Restless==0.17.0
python-decouple==2.4
pytest
psycopg2

